requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic[email]==2.6.1
sqlalchemy==2.0.27
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
bcrypt==4.3.0